Move CliFactory to cmd package
